ml-model/requirements.txt
torch
torchvision
fastapi
uvicorn
python-multipart
Pillow
scikit-learn
numpy
pandas
python-multipart